Add `gc-content-plot' program: src/Align/gc-content-plot.cc
This program is similar to the kmer-cov-plot program, but instead of printing
k-mer coverage along each read, contig, or scaffold, it will print the GC
content within a sliding window of configurable size.  The GC content can be
printed as an absolute number or as a percentage.  Sequences containing unknown
DNA characters can either be excluded, or unknown DNA characters can be assigned
a default GC value.  A tabular output format is supported.