alignextend.pl changes
Changes to the alignextend.pl script:

- You can now save the unaligned pairs to FASTQ files by providing the
'-unaligned' option.
- The informational output is now more clear about exactly what fraction of
reads aligned, didn't align, aligned to the same scaffold, etc.
- The preparation of the input FASTQ files has been moved out of the "align"
stage and into a new stage, "prepare".
- sam2bed() and matchbed() are combined into a function filter_by_alignment() in
a stage now renamed to "filter".  It's assumed that the reads are already paired
up properly and that BWA preserves the pairing.
- You can now specify the '-n' option to BWA with the option '-maxdiff'.
- You can now skip the "rmdup" stage by providing the "-skiprmdup" option.
- Quality values of the extended reads are now set to the expected value rather
than always to 'a' or 'A'.  Default is now 31 plus the quality offset.
- Check for errors when reading, writing, and closing files.
- The flags in the SAM flags field are now tested bitwise.
- Timing information is printed after each stage has completed.